refactor tings
