+++ src/components/shared/VolumeChart/daysToMillis.ts
@@ -1,0 +1,3 @@
+import { SECONDS_IN_DAY } from 'constants/numbers'
+
+export const daysToMillis = (days: number) => days * SECONDS_IN_DAY * 1000
